+++ src/CMakeLists.txt
@@ -348,11 +348,32 @@ set(SOURCE_FILES
     WrapCalls.cpp
     )
 
+set(TEMPLATE_FILES
+    CodeGen_C_prologue
+    CodeGen_C_vectors
+    )
+
 ##
 # Build and import the runtime.
 ##
 
 add_subdirectory(runtime)
+
+##
+# Build the template files via binary2cpp.
+##
+
+add_library(Halide_c_templates OBJECT)
+foreach (f IN LISTS TEMPLATE_FILES)
+    set(SRC "$<SHELL_PATH:${CMAKE_CURRENT_SOURCE_DIR}/${f}.template.cpp>")
+    set(DST "c_template.${f}.template.cpp")
+
+    add_custom_command(OUTPUT "${DST}"
+                       COMMAND binary2cpp "halide_c_template_${f}" < "${SRC}" > "${DST}"
+                       DEPENDS "${SRC}" binary2cpp
+                       VERBATIM)
+    target_sources(Halide_c_templates PRIVATE ${DST})
+endforeach ()
 
 ##
 # Build the Halide mono-header.
@@ -376,7 +397,8 @@ add_library(Halide
             ${SOURCE_FILES}
             ${HEADER_FILES}
             # Including these as sources works around the need to "install" Halide_initmod
-            $<TARGET_OBJECTS:Halide_initmod>)
+            $<TARGET_OBJECTS:Halide_initmod>
+            $<TARGET_OBJECTS:Halide_c_templates>)
 add_library(Halide::Halide ALIAS Halide)
 
 target_link_libraries(Halide PRIVATE Halide::LLVM)
