+++ src/CodeGen_C.cpp
@@ -34,6 +34,8 @@ extern "C" unsigned char halide_internal_runtime_header_HalideRuntimeOpenGLCompu
 extern "C" unsigned char halide_internal_runtime_header_HalideRuntimeQurt_h[];
 extern "C" unsigned char halide_internal_runtime_header_HalideRuntimeD3D12Compute_h[];
 extern "C" unsigned char halide_internal_runtime_header_HalideRuntimeWebGPU_h[];
+extern "C" unsigned char halide_c_template_CodeGen_C_prologue[];
+extern "C" unsigned char halide_c_template_CodeGen_C_vectors[];
 
 namespace {
 
@@ -53,197 +55,6 @@ const char *const kDefineMustUseResult = R"INLINE_CODE(#ifndef HALIDE_MUST_USE_R
 #define HALIDE_MUST_USE_RESULT
 #endif
 #endif
-)INLINE_CODE";
-
-const string headers = R"INLINE_CODE(
-/* MACHINE GENERATED By Halide. */
-
-#if !(__cplusplus >= 201103L || _MSVC_LANG >= 201103L)
-#error "This code requires C++11 (or later); please upgrade your compiler."
-#endif
-
-#include <assert.h>
-#include <float.h>
-#include <limits.h>
-#include <math.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <string.h>
-#include <type_traits>
-#include <fenv.h>
-)INLINE_CODE";
-
-// We now add definitions of things in the runtime which are
-// intended to be inlined into every module but are only expressed
-// in .ll. The redundancy is regrettable (FIXME).
-const string globals = R"INLINE_CODE(
-extern "C" {
-int64_t halide_current_time_ns(void *ctx);
-void halide_profiler_pipeline_end(void *, void *);
-struct halide_buffer_t;
-char *halide_buffer_to_string(char *, char *, const halide_buffer_t *);
-}
-
-#ifdef _WIN32
-#ifndef _MT
-__declspec(dllimport) float __cdecl roundf(float);
-__declspec(dllimport) double __cdecl round(double);
-#endif
-#else
-inline float asinh_f32(float x) {return asinhf(x);}
-inline float acosh_f32(float x) {return acoshf(x);}
-inline float atanh_f32(float x) {return atanhf(x);}
-inline double asinh_f64(double x) {return asinh(x);}
-inline double acosh_f64(double x) {return acosh(x);}
-inline double atanh_f64(double x) {return atanh(x);}
-#endif
-inline float sqrt_f32(float x) {return sqrtf(x);}
-inline float sin_f32(float x) {return sinf(x);}
-inline float asin_f32(float x) {return asinf(x);}
-inline float cos_f32(float x) {return cosf(x);}
-inline float acos_f32(float x) {return acosf(x);}
-inline float tan_f32(float x) {return tanf(x);}
-inline float atan_f32(float x) {return atanf(x);}
-inline float atan2_f32(float x, float y) {return atan2f(x, y);}
-inline float sinh_f32(float x) {return sinhf(x);}
-inline float cosh_f32(float x) {return coshf(x);}
-inline float tanh_f32(float x) {return tanhf(x);}
-inline float hypot_f32(float x, float y) {return hypotf(x, y);}
-inline float exp_f32(float x) {return expf(x);}
-inline float log_f32(float x) {return logf(x);}
-inline float pow_f32(float x, float y) {return powf(x, y);}
-inline float floor_f32(float x) {return floorf(x);}
-inline float ceil_f32(float x) {return ceilf(x);}
-
-inline double sqrt_f64(double x) {return sqrt(x);}
-inline double sin_f64(double x) {return sin(x);}
-inline double asin_f64(double x) {return asin(x);}
-inline double cos_f64(double x) {return cos(x);}
-inline double acos_f64(double x) {return acos(x);}
-inline double tan_f64(double x) {return tan(x);}
-inline double atan_f64(double x) {return atan(x);}
-inline double atan2_f64(double x, double y) {return atan2(x, y);}
-inline double sinh_f64(double x) {return sinh(x);}
-inline double cosh_f64(double x) {return cosh(x);}
-inline double tanh_f64(double x) {return tanh(x);}
-inline double hypot_f64(double x, double y) {return hypot(x, y);}
-inline double exp_f64(double x) {return exp(x);}
-inline double log_f64(double x) {return log(x);}
-inline double pow_f64(double x, double y) {return pow(x, y);}
-inline double floor_f64(double x) {return floor(x);}
-inline double ceil_f64(double x) {return ceil(x);}
-
-inline float nan_f32() {return NAN;}
-inline float neg_inf_f32() {return -INFINITY;}
-inline float inf_f32() {return INFINITY;}
-inline bool is_nan_f32(float x) {return isnan(x);}
-inline bool is_nan_f64(double x) {return isnan(x);}
-inline bool is_inf_f32(float x) {return isinf(x);}
-inline bool is_inf_f64(double x) {return isinf(x);}
-inline bool is_finite_f32(float x) {return isfinite(x);}
-inline bool is_finite_f64(double x) {return isfinite(x);}
-
-template<typename A, typename B>
-inline A reinterpret(const B &b) {
-    static_assert(sizeof(A) == sizeof(B), "type size mismatch");
-    A a;
-    memcpy(&a, &b, sizeof(a));
-    return a;
-}
-inline float float_from_bits(uint32_t bits) {
-    return reinterpret<float, uint32_t>(bits);
-}
-
-template<typename T>
-inline int halide_popcount_fallback(T a) {
-    int bits_set = 0;
-    while (a != 0) {
-        bits_set += 1;
-        // this is Brian Kernigan's bit counting algorithm
-        // https://graphics.stanford.edu/~seander/bithacks.html#CountBitsSetKernighan)
-        // removes the least significant one
-        a &= a - 1;
-    }
-    return bits_set;
-}
-
-template<typename T>
-inline int halide_popcount(T a) {
-    return halide_popcount_fallback<T>(a);
-}
-
-template<>
-inline int halide_popcount<uint64_t>(uint64_t a) {
-#ifdef _MSC_VER
-#if defined(_WIN64)
-    return __popcnt64(a);
-#else
-    return __popcnt((uint32_t)(a >> 32)) + __popcnt((uint32_t)(a & 0xffffffff));
-#endif
-#else
-#if defined(__builtin_popcountll)
-    static_assert(sizeof(unsigned long long) >= sizeof(uint64_t), "");
-    return return __builtin_popcountll(a);
-#else
-    return halide_popcount_fallback<uint64_t>(a);
-#endif
-#endif
-}
-
-template<typename T>
-inline int halide_count_leading_zeros(T a) {
-    int leading_zeros = 0;
-    int bit = sizeof(a) * 8 - 1;
-    while (bit >= 0 && (a & (((T)1) << bit)) == 0) {
-        leading_zeros++;
-        bit--;
-    }
-    return leading_zeros;
-}
-
-template<typename T>
-inline int halide_count_trailing_zeros(T a) {
-    int trailing_zeros = 0;
-    constexpr int bits = sizeof(a) * 8;
-    int bit = 0;
-    while (bit < bits && (a & (((T)1) << bit)) == 0) {
-        trailing_zeros++;
-        bit++;
-    }
-    return trailing_zeros;
-}
-
-template<typename T>
-inline T halide_cpp_max(const T &a, const T &b) {return (a > b) ? a : b;}
-
-template<typename T>
-inline T halide_cpp_min(const T &a, const T &b) {return (a < b) ? a : b;}
-
-template<typename T>
-inline void halide_maybe_unused(const T&) {}
-
-template<typename A, typename B>
-const B &return_second(const A &a, const B &b) {
-    halide_maybe_unused(a);
-    return b;
-}
-
-namespace {
-template<void(*FreeFn)(void *, void *)>
-class HalideFreeHelper {
-    void * const user_context;
-    void * ptr;
-public:
-    HalideFreeHelper(void *user_context, void *ptr) : user_context(user_context), ptr(ptr) {}
-    ~HalideFreeHelper() { free(); }
-    void free() {
-        if (ptr) {
-            FreeFn(user_context, ptr);
-            ptr = nullptr;
-        }
-    }
-};
-} // namespace
 )INLINE_CODE";
 
 const char *const constexpr_argument_info_docs = R"INLINE_CODE(
@@ -438,8 +249,7 @@ CodeGen_C::CodeGen_C(ostream &s, const Target &t, OutputKind output_kind, const 
     } else {
         // Include declarations of everything generated C source might want
         stream
-            << headers
-            << globals
+            << halide_c_template_CodeGen_C_prologue << "\n"
             << halide_internal_runtime_header_HalideRuntime_h << "\n"
             << halide_internal_initmod_inlined_c << "\n";
         stream << "\n";
@@ -519,819 +329,12 @@ void CodeGen_C::add_platform_prologue() {
 
 void CodeGen_C::add_vector_typedefs(const std::set<Type> &vector_types) {
     if (!vector_types.empty()) {
-        // MSVC has a limit of ~16k for string literals, so split
-        // up these declarations accordingly
-        const char *cpp_vector_decl = R"INLINE_CODE(
-#if !defined(__has_attribute)
-    #define __has_attribute(x) 0
-#endif
-
-#if !defined(__has_builtin)
-    #define __has_builtin(x) 0
-#endif
-
-namespace {
-
-// We can't use std::array because that has its own overload of operator<, etc,
-// which will interfere with ours.
-template <typename ElementType, size_t Lanes>
-struct CppVector {
-    ElementType elements[Lanes];
-
-    HALIDE_ALWAYS_INLINE
-    ElementType& operator[](size_t i) {
-        return elements[i];
-    }
-
-    HALIDE_ALWAYS_INLINE
-    const ElementType operator[](size_t i) const {
-        return elements[i];
-    }
-
-    HALIDE_ALWAYS_INLINE
-    ElementType *data() {
-        return elements;
-    }
-
-    HALIDE_ALWAYS_INLINE
-    const ElementType *data() const {
-        return elements;
-    }
-};
-
-template <typename ElementType_, size_t Lanes_>
-class CppVectorOps {
-public:
-    using ElementType = ElementType_ ;
-    static constexpr size_t Lanes = Lanes_;
-
-    using Vec = CppVector<ElementType, Lanes>;
-    using Mask = CppVector<uint8_t, Lanes>;
-
-    CppVectorOps() = delete;
-
-    static Vec broadcast(const ElementType v) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = v;
-        }
-        return r;
-    }
-
-    static Vec ramp(const ElementType base, const ElementType stride) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = base + stride * i;
-        }
-        return r;
-    }
-
-    static Vec load(const void *base, int32_t offset) {
-        Vec r;
-        memcpy(r.data(), ((const ElementType*)base + offset), sizeof(ElementType) * Lanes);
-        return r;
-    }
-
-    static Vec load_gather(const void *base, const CppVector<int32_t, Lanes> &offset) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = ((const ElementType*)base)[offset[i]];
-        }
-        return r;
-    }
-
-    static Vec load_predicated(const void *base, const CppVector<int32_t, Lanes> &offset, const Mask &predicate) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            if (predicate[i]) {
-                r[i] = ((const ElementType*)base)[offset[i]];
-            }
-        }
-        return r;
-    }
-
-    static void store(const Vec &v, void *base, int32_t offset) {
-        memcpy(((ElementType*)base + offset), v.data(), sizeof(ElementType) * Lanes);
-    }
-
-    static void store_scatter(const Vec &v, void *base, const CppVector<int32_t, Lanes> &offset) {
-        for (size_t i = 0; i < Lanes; i++) {
-            ((ElementType*)base)[offset[i]] = v[i];
-        }
-    }
-
-    static void store_predicated(const Vec &v, void *base, const CppVector<int32_t, Lanes> &offset, const Mask &predicate) {
-        for (size_t i = 0; i < Lanes; i++) {
-            if (predicate[i]) {
-                ((ElementType*)base)[offset[i]] = v[i];
-            }
-        }
-    }
-
-    template<int... Indices, typename InputVec>
-    static Vec shuffle(const InputVec &a) {
-        static_assert(sizeof...(Indices) == Lanes, "shuffle() requires an exact match of lanes");
-        Vec r = { a[Indices]... };
-        return r;
-    }
-
-    static Vec replace(const Vec &v, size_t i, const ElementType b) {
-        Vec r = v;
-        r[i] = b;
-        return r;
-    }
-
-    template <typename OtherVec>
-    static Vec convert_from(const OtherVec &src) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = static_cast<ElementType>(src[i]);
-        }
-        return r;
-    }
-
-    static Vec max(const Vec &a, const Vec &b) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = ::halide_cpp_max(a[i], b[i]);
-        }
-        return r;
-    }
-
-    static Vec min(const Vec &a, const Vec &b) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = ::halide_cpp_min(a[i], b[i]);
-        }
-        return r;
-    }
-
-    static Vec select(const Mask &cond, const Vec &true_value, const Vec &false_value) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = cond[i] ? true_value[i] : false_value[i];
-        }
-        return r;
-    }
-
-    static Mask logical_or(const Vec &a, const Vec &b) {
-        CppVector<uint8_t, Lanes> r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] || b[i] ? 0xff : 0x00;
-        }
-        return r;
-    }
-
-    static Mask logical_and(const Vec &a, const Vec &b) {
-        CppVector<uint8_t, Lanes> r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] && b[i] ? 0xff : 0x00;
-        }
-        return r;
-    }
-
-    static Mask lt(const Vec &a, const Vec &b) {
-        CppVector<uint8_t, Lanes> r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] < b[i] ? 0xff : 0x00;
-        }
-        return r;
-    }
-
-    static Mask le(const Vec &a, const Vec &b) {
-        CppVector<uint8_t, Lanes> r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] <= b[i] ? 0xff : 0x00;
-        }
-        return r;
-    }
-
-    static Mask gt(const Vec &a, const Vec &b) {
-        CppVector<uint8_t, Lanes> r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] > b[i] ? 0xff : 0x00;
-        }
-        return r;
-    }
-
-    static Mask ge(const Vec &a, const Vec &b) {
-        CppVector<uint8_t, Lanes> r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] >= b[i] ? 0xff : 0x00;
-        }
-        return r;
-    }
-
-    static Mask eq(const Vec &a, const Vec &b) {
-        CppVector<uint8_t, Lanes> r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] == b[i] ? 0xff : 0x00;
-        }
-        return r;
-    }
-
-    static Mask ne(const Vec &a, const Vec &b) {
-        CppVector<uint8_t, Lanes> r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] != b[i] ? 0xff : 0x00;
-        }
-        return r;
-    }
-};
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator~(const CppVector<ElementType, Lanes> &v) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = ~v[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator!(const CppVector<ElementType, Lanes> &v) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = !v[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator+(const CppVector<ElementType, Lanes> &a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] + b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator-(const CppVector<ElementType, Lanes> &a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] - b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator*(const CppVector<ElementType, Lanes> &a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] * b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator/(const CppVector<ElementType, Lanes> &a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] / b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator%(const CppVector<ElementType, Lanes> &a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] % b[i];
-    }
-    return r;
-}
-
-template <typename ElementType, size_t Lanes, typename OtherElementType>
-CppVector<ElementType, Lanes> operator<<(const CppVector<ElementType, Lanes> &a, const CppVector<OtherElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] << b[i];
-    }
-    return r;
-}
-
-template <typename ElementType, size_t Lanes, typename OtherElementType>
-CppVector<ElementType, Lanes> operator>>(const CppVector<ElementType, Lanes> &a, const CppVector<OtherElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] >> b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator&(const CppVector<ElementType, Lanes> &a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] & b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator|(const CppVector<ElementType, Lanes> &a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] | b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator^(const CppVector<ElementType, Lanes> &a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] ^ b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator+(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] + b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator-(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] - b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator*(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] * b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator/(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] / b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator%(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] % b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator>>(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] >> b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator<<(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] << b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator&(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] & b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator|(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] | b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator^(const CppVector<ElementType, Lanes> &a, const ElementType b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a[i] ^ b;
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator+(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a + b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator-(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a - b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator*(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a * b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator/(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a / b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator%(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a % b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator>>(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a >> b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator<<(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a << b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator&(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a & b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator|(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a | b[i];
-    }
-    return r;
-}
-
-template<typename ElementType, size_t Lanes>
-CppVector<ElementType, Lanes> operator^(const ElementType a, const CppVector<ElementType, Lanes> &b) {
-    CppVector<ElementType, Lanes> r;
-    for (size_t i = 0; i < Lanes; i++) {
-        r[i] = a ^ b[i];
-    }
-    return r;
-}
-
-}  // namespace
-
-
-)INLINE_CODE";
-
-        const char *native_vector_decl = R"INLINE_CODE(
-namespace {
-
-#if __has_attribute(ext_vector_type) || __has_attribute(vector_size)
-
-#if __has_attribute(ext_vector_type)
-    // Clang
-    template<typename ElementType, size_t Lanes>
-    using NativeVector __attribute__((ext_vector_type(Lanes), aligned(sizeof(ElementType)))) = ElementType;
-#elif __has_attribute(vector_size) || defined(__GNUC__)
-    // GCC
-    template<typename ElementType, size_t Lanes>
-    using NativeVector __attribute__((vector_size(Lanes * sizeof(ElementType)), aligned(sizeof(ElementType)))) = ElementType;
-#else
-    #error
-#endif
-
-template<typename T>
-struct NativeVectorComparisonType {
-    using type = void;
-};
-
-template<>
-struct NativeVectorComparisonType<int8_t> { using type = char; };
-
-template<>
-struct NativeVectorComparisonType<int16_t> { using type = int16_t; };
-
-template<>
-struct NativeVectorComparisonType<int32_t> { using type = int32_t; };
-
-template<>
-struct NativeVectorComparisonType<int64_t> { using type = int64_t; };
-
-template<>
-struct NativeVectorComparisonType<uint8_t> { using type = char; };
-
-template<>
-struct NativeVectorComparisonType<uint16_t> { using type = int16_t; };
-
-template<>
-struct NativeVectorComparisonType<uint32_t> { using type = int32_t; };
-
-template<>
-struct NativeVectorComparisonType<uint64_t> { using type = int64_t; };
-
-template<>
-struct NativeVectorComparisonType<float> { using type = int32_t; };
-
-template<>
-struct NativeVectorComparisonType<double> { using type = int64_t; };
-
-template <typename ElementType_, size_t Lanes_>
-class NativeVectorOps {
-public:
-    using ElementType = ElementType_ ;
-    static constexpr size_t Lanes = Lanes_;
-
-    using Vec = NativeVector<ElementType, Lanes>;
-    using Mask = NativeVector<uint8_t, Lanes>;
-
-    NativeVectorOps() = delete;
-
-    static Vec broadcast(const ElementType v) {
-        const Vec zero = {}; // Zero-initialized native vector.
-        return v - zero;
-    }
-
-    static Vec ramp(const ElementType base, const ElementType stride) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = base + stride * i;
-        }
-        return r;
-    }
-
-    static Vec load(const void *base, int32_t offset) {
-        Vec r;
-        // We only require Vec to be element-aligned, so we can't safely just read
-        // directly from memory (might segfault). Use memcpy for safety.
-        //
-        // If Vec is a non-power-of-two (e.g. uint8x48), the actual implementation
-        // might be larger (e.g. it might really be a uint8x64). Only copy the amount
-        // that is in the logical type, to avoid possible overreads.
-        memcpy(&r, ((const ElementType*)base + offset), sizeof(ElementType) * Lanes);
-        return r;
-    }
-
-    static Vec load_gather(const void *base, const NativeVector<int32_t, Lanes> offset) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = ((const ElementType*)base)[offset[i]];
-        }
-        return r;
-    }
-
-    static Vec load_predicated(const void *base, const NativeVector<int32_t, Lanes> offset, const NativeVector<uint8_t, Lanes> predicate) {
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            if (predicate[i]) {
-                r[i] = ((const ElementType*)base)[offset[i]];
-            }
-        }
-        return r;
-    }
-    static void store(const Vec v, void *base, int32_t offset) {
-        // We only require Vec to be element-aligned, so we can't safely just write
-        // directly from memory (might segfault). Use memcpy for safety.
-        //
-        // If Vec is a non-power-of-two (e.g. uint8x48), the actual implementation
-        // might be larger (e.g. it might really be a uint8x64). Only copy the amount
-        // that is in the logical type, to avoid possible overreads.
-        memcpy(((ElementType*)base + offset), &v, sizeof(ElementType) * Lanes);
-    }
-
-    static void store_scatter(const Vec v, void *base, const NativeVector<int32_t, Lanes> offset) {
-        for (size_t i = 0; i < Lanes; i++) {
-            ((ElementType*)base)[offset[i]] = v[i];
-        }
-    }
-
-    static void store_predicated(const Vec v, void *base, const NativeVector<int32_t, Lanes> offset, const NativeVector<uint8_t, Lanes> predicate) {
-        for (size_t i = 0; i < Lanes; i++) {
-            if (predicate[i]) {
-                ((ElementType*)base)[offset[i]] = v[i];
-            }
-        }
-    }
-
-    template<int... Indices, typename InputVec>
-    static Vec shuffle(const InputVec a) {
-        static_assert(sizeof...(Indices) == Lanes, "shuffle() requires an exact match of lanes");
-#if __has_builtin(__builtin_shufflevector)
-        // Exists in clang and gcc >= 12. Gcc's __builtin_shuffle can't
-        // be used, because it can't handle changing the number of vector
-        // lanes between input and output.
-        return __builtin_shufflevector(a, a, Indices...);
-#else
-        Vec r = { a[Indices]... };
-        return r;
-#endif
-    }
-
-    static Vec replace(Vec v, size_t i, const ElementType b) {
-        v[i] = b;
-        return v;
-    }
-
-    template <typename OtherVec>
-    static Vec convert_from(const OtherVec src) {
-#if __has_builtin(__builtin_convertvector)
-        // Don't use __builtin_convertvector for float->int: it appears to have
-        // different float->int rounding behavior in at least some situations;
-        // for now we'll use the much-slower-but-correct explicit C++ code.
-        // (https://github.com/halide/Halide/issues/2080)
-        constexpr bool is_float_to_int = std::is_floating_point<OtherVec>::value &&
-                                         std::is_integral<Vec>::value;
-        if (!is_float_to_int) {
-            return __builtin_convertvector(src, Vec);
-        }
-#endif
-        // Fallthru for float->int, or degenerate compilers that support native vectors
-        // but not __builtin_convertvector (Intel?)
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = static_cast<ElementType>(src[i]);
-        }
-        return r;
-    }
-
-    static Vec max(const Vec a, const Vec b) {
-#if defined(__GNUC__) && !defined(__clang__)
-        // TODO: GCC doesn't seem to recognize this pattern, and scalarizes instead
-        return a > b ? a : b;
-#else
-        // Clang doesn't do ternary operator for vectors, but recognizes this pattern
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] > b[i] ? a[i] : b[i];
-        }
-        return r;
-#endif
-    }
-
-    static Vec min(const Vec a, const Vec b) {
-#if defined(__GNUC__) && !defined(__clang__)
-        // TODO: GCC doesn't seem to recognize this pattern, and scalarizes instead
-        return a < b ? a : b;
-#else
-        // Clang doesn't do ternary operator for vectors, but recognizes this pattern
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = a[i] < b[i] ? a[i] : b[i];
-        }
-        return r;
-#endif
-    }
-
-    static Vec select(const Mask cond, const Vec true_value, const Vec false_value) {
-#if defined(__GNUC__) && !defined(__clang__)
-        // This should do the correct lane-wise select.
-        using T = typename NativeVectorComparisonType<ElementType>::type;
-        auto b = NativeVectorOps<T, Lanes>::convert_from(cond);
-        return b ? true_value : false_value;
-#else
-        // Clang doesn't do ternary operator for vectors, but recognizes this pattern
-        Vec r;
-        for (size_t i = 0; i < Lanes; i++) {
-            r[i] = cond[i] ? true_value[i] : false_value[i];
-        }
-        return r;
-#endif
-    }
-
-    // The relational operators produce signed-int of same width as input; our codegen expects uint8.
-    static Mask logical_or(const Vec a, const Vec b) {
-        using T = typename NativeVectorComparisonType<ElementType>::type;
-        const NativeVector<T, Lanes> r = a || b;
-        return NativeVectorOps<uint8_t, Lanes>::convert_from(r);
-    }
-
-    static Mask logical_and(const Vec a, const Vec b) {
-        using T = typename NativeVectorComparisonType<ElementType>::type;
-        const NativeVector<T, Lanes> r = a && b;
-        return NativeVectorOps<uint8_t, Lanes>::convert_from(r);
-    }
-
-    static Mask lt(const Vec a, const Vec b) {
-        using T = typename NativeVectorComparisonType<ElementType>::type;
-        const NativeVector<T, Lanes> r = a < b;
-        return NativeVectorOps<uint8_t, Lanes>::convert_from(r);
-    }
-
-    static Mask le(const Vec a, const Vec b) {
-        using T = typename NativeVectorComparisonType<ElementType>::type;
-        const NativeVector<T, Lanes> r = a <= b;
-        return NativeVectorOps<uint8_t, Lanes>::convert_from(r);
-    }
-
-    static Mask gt(const Vec a, const Vec b) {
-        using T = typename NativeVectorComparisonType<ElementType>::type;
-        const NativeVector<T, Lanes> r = a > b;
-        return NativeVectorOps<uint8_t, Lanes>::convert_from(r);
-    }
-
-    static Mask ge(const Vec a, const Vec b) {
-        using T = typename NativeVectorComparisonType<ElementType>::type;
-        const NativeVector<T, Lanes> r = a >= b;
-        return NativeVectorOps<uint8_t, Lanes>::convert_from(r);
-    }
-
-    static Mask eq(const Vec a, const Vec b) {
-        using T = typename NativeVectorComparisonType<ElementType>::type;
-        const NativeVector<T, Lanes> r = a == b;
-        return NativeVectorOps<uint8_t, Lanes>::convert_from(r);
-    }
-
-    static Mask ne(const Vec a, const Vec b) {
-        using T = typename NativeVectorComparisonType<ElementType>::type;
-        const NativeVector<T, Lanes> r = a != b;
-        return NativeVectorOps<uint8_t, Lanes>::convert_from(r);
-    }
-};
-
-
-#endif  // __has_attribute(ext_vector_type) || __has_attribute(vector_size)
-
-}  // namespace
-
-)INLINE_CODE";
-
-        const char *vector_selection_decl = R"INLINE_CODE(
-// Dec. 1, 2018: Apparently emscripten compilation runs with the __has_attribute true,
-// then fails to handle the vector intrinsics later.
-#if !defined(__EMSCRIPTEN__) && (__has_attribute(ext_vector_type) || __has_attribute(vector_size))
-    #if __GNUC__ && !__clang__
-        // GCC only allows powers-of-two; fall back to CppVector for other widths
-        #define halide_cpp_use_native_vector(type, lanes) ((lanes & (lanes - 1)) == 0)
-    #else
-        #define halide_cpp_use_native_vector(type, lanes) (true)
-    #endif
-#else
-    // No NativeVector available
-    #define halide_cpp_use_native_vector(type, lanes) (false)
-#endif  // __has_attribute(ext_vector_type) || __has_attribute(vector_size)
-
-// Failsafe to allow forcing non-native vectors in case of unruly compilers
-#if HALIDE_CPP_ALWAYS_USE_CPP_VECTORS
-    #undef halide_cpp_use_native_vector
-    #define halide_cpp_use_native_vector(type, lanes) (false)
-#endif
-
-)INLINE_CODE";
-
-        // Vodoo fix: on at least one config (our arm32 buildbot running gcc 5.4),
+        // Voodoo fix: on at least one config (our arm32 buildbot running gcc 5.4),
         // emitting this long text string was regularly garbled in a predictable pattern;
         // flushing the stream before or after heals it. Since C++ codegen is rarely
         // on a compilation critical path, we'll just band-aid it in this way.
         stream << std::flush;
-        stream << cpp_vector_decl << native_vector_decl << vector_selection_decl;
+        stream << halide_c_template_CodeGen_C_vectors;
         stream << std::flush;
 
         for (const auto &t : vector_types) {
@@ -1501,7 +504,7 @@ class ExternCallPrototypes : public IRGraphVisitor {
 public:
     ExternCallPrototypes() {
         // Make sure we don't catch calls that are already in the global declarations
-        const char *strs[] = {globals.c_str(),
+        const char *strs[] = {(const char *)halide_c_template_CodeGen_C_prologue,
                               (const char *)halide_internal_runtime_header_HalideRuntime_h,
                               (const char *)halide_internal_initmod_inlined_c};
         for (const char *str : strs) {
@@ -3531,8 +2534,7 @@ void CodeGen_C::test() {
     }
 
     string correct_source =
-        headers +
-        globals +
+        string((const char *)halide_c_template_CodeGen_C_prologue) + '\n' +
         string((const char *)halide_internal_runtime_header_HalideRuntime_h) + '\n' +
         string((const char *)halide_internal_initmod_inlined_c) + '\n' +
         '\n' + kDefineMustUseResult + R"GOLDEN_CODE(
